requirements.txt
bcrypt==5.0.0

blinker==1.9.0

click==8.3.1

colorama==0.4.6

Flask==3.1.2

flask-cors==6.0.2

Flask-JWT-Extended==4.7.1

greenlet==3.3.1

itsdangerous==2.2.0

Jinja2==3.1.6

MarkupSafe==3.0.3

numpy==2.4.2

psycopg==3.2.4

PyJWT==2.11.0

python-dotenv==1.2.1

SQLAlchemy==2.0.46

typing_extensions==4.15.0

Werkzeug==3.1.5

